# assets/openapi.yaml
openapi: 3.1.0
info:
  title: openccu-loom REST API
  version: 1.0.0
  description: |
    North-bound REST + WebSocket surface of the openccu-loom daemon.
    Errors follow RFC 9457 problem+json.
  license:
    name: MIT
    url: https://opensource.org/license/mit/
servers:
  - url: /api/v1
security:
  - basicAuth: []
  - bearerAuth: []
  - openIdConnect: [viewer]
...
paths:
...
  /devices/values:batch:
    post:
      operationId: valuesBatch
      summary: Batch-read data-point values in one round trip
      description: |
        Returns the requested data-point summaries in a single
        request — avoids the N-roundtrips initial-sync penalty
        external clients (HA, dashboards, monitoring) hit when
        loading 80k DPs over per-DP GETs.

        Per-query errors (missing device, channel, parameter, or
        visibility-hidden parameter) surface inside the response
        body so one bad address does not abort the rest of the
        batch. Envelope-level problems (malformed JSON, empty
        queries, exceeding the per-request cap) yield problem+json.

        The per-request query cap is currently 1000.
      tags: [devices]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ValuesBatchRequest" }
      responses:
        "200":
          description: Per-query results (success or error)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ValuesBatchResponse" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          description: Empty queries / oversized batch
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
  /diagnostics:
    get:
      operationId: diagnosticsDump
      summary: Composite diagnostics dump (health + interfaces + incidents + log-levels)
      description: |
        Single ship-the-state artefact for support / agent escalation
        (ADR 0017). One JSON envelope carrying build info, health
        components + score, per-client detail, configured-interface
        list, recent incidents, system-status ring buffer, and
        log-level overrides.

        **Two consumer planes — disjoint:**

        - `GET /diagnostics` returns a momentary structured JSON
          snapshot. Right for support escalations, HA Repair-Flows,
          and any client that needs a coherent point-in-time view
          with one call.
        - `GET /metrics` returns Prometheus exposition format,
          right for time-series scraping (Grafana, Alertmanager).
          Continuous monitoring uses this; diagnostics-on-demand
          uses the JSON endpoint above.

        The two surfaces are not interchangeable: the JSON envelope
        carries fields (recent incidents, log-overrides, system-status
        ring) that have no useful Prometheus representation; the
        Prometheus surface emits time-series that the JSON envelope
        does not aggregate.

        Hashes device-address-shaped fields with a 12-hex sha256
        prefix by default; pass `?anonymize=0` explicitly when
        absolute fidelity is required.
      tags: [diagnostics]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [admin]
      parameters:
        - in: query
          name: anonymize
          schema: { type: string, enum: ["0", "1"], default: "1" }
      responses:
        "200":
          description: Diagnostics envelope
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /diagnostics/log-levels:
    get:
      operationId: listLogLevels
      summary: Active per-subsystem log-level overrides
      tags: [diagnostics]
      responses:
        "200":
          description: Override list
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /diagnostics/log-levels/{path}:
    put:
      operationId: putLogLevel
      summary: Install a per-subsystem log-level override (admin-only)
      tags: [diagnostics]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [admin]
      parameters:
        - in: path
          name: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "204":
          description: Override applied
        "400": { $ref: "#/components/responses/BadRequest" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
    delete:
      operationId: deleteLogLevel
      summary: Remove a per-subsystem log-level override (admin-only)
      tags: [diagnostics]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [admin]
      parameters:
        - in: path
          name: path
          required: true
          schema: { type: string }
      responses:
        "204": { description: Removed }
        "400": { $ref: "#/components/responses/BadRequest" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /diagnostics/log-level:
    get:
      operationId: getDefaultLogLevel
      summary: Current global default log level
      tags: [diagnostics]
      responses:
        "200":
          description: Default level
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
    put:
      operationId: putDefaultLogLevel
      summary: Change the global default log level (admin-only)
      tags: [diagnostics]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [admin]
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, additionalProperties: true }
      responses:
        "200":
          description: Default level updated
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        "400": { $ref: "#/components/responses/BadRequest" }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
  /diagnostics/logs:
    get:
      operationId: listLogs
      summary: Backfill / download recent log records (admin-only)
      tags: [diagnostics]
      security:
        - basicAuth: []
        - bearerAuth: []
        - openIdConnect: [admin]
      parameters:
        - in: query
          name: limit
          schema: { type: integer }
          description: Max records (default 500, capped at the ring size).
        - in: query
          name: min_level
          schema: { type: string, enum: [debug, info, warn, error] }
        - in: query
          name: format
          schema: { type: string, enum: [json, ndjson] }
        - in: query
          name: download
          schema: { type: string, enum: ["1"] }
          description: When "1", sets a Content-Disposition attachment header.
      responses:
        "200":
          description: Recent log records
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
            application/x-ndjson:
              schema: { type: string }
        "503": { $ref: "#/components/responses/ServiceUnavailable" }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: |
        Username + password (operator-grade or above for mutations).
        The role hierarchy (viewer ⊂ operator ⊂ admin) is enforced
        server-side regardless of authentication scheme.
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API token issued via the auth tokens surface (admin-only
        provisioning). Token's stored role determines authorisation.
    openIdConnect:
      type: openIdConnect
      description: |
        OpenID Connect authentication. The operator configures the
        OIDC discovery URL; the placeholder below is the standard
        well-known path. Per-operation `security:` blocks express
        the **minimum role** the daemon enforces via OIDC-style
        scopes — the same enforcement applies when the client
        authenticates via basicAuth or bearerAuth, but those schemes
        cannot natively express scopes in OpenAPI 3.x.

        Scope vocabulary mirrors `Identity.role`:
          - `viewer` — read-only operations
          - `operator` — mutations (write data points, manage links,
            edit schedules, run config-import, trigger refresh)
          - `admin` — dangerous operations (delete device, edit log
            levels, list users/tokens, trigger backup, manage
            interfaces, edit Matter fabrics)
      openIdConnectUrl: /.well-known/openid-configuration
...
  responses:
    BadRequest:
      description: Validation or parsing failure
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    ServiceUnavailable:
      description: Backend not wired
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
  schemas:
...
    Problem:
      type: object
      description: |
        RFC 9457 problem detail. The `type` URI selects one of the
        canonical problem kinds the daemon emits; the `code` member
        carries the short tag (also surfaced as the `X-Problem-Code`
        response header). Clients should switch on `code` for typed
        error handling — the `type` URI is the human-readable form.

        **Rate-limiting:** when the daemon emits `code: rate_limited`
        (HTTP 429) it sets the `Retry-After` response header per
        RFC 9110 §10.2.3 with the integer number of seconds the
        client should wait before retrying. Clients that respect
        the header are immune to the throttle; clients that don't
        will see repeated 429s with the same `Retry-After` value
        until the throttle window expires.
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: |
            Stable URI for one of the canonical problem kinds. All
            URIs share the base `https://openccu-loom.dev/errors/`
            and the trailing segment matches the `code` enum below.
            The base URI is a stable identifier, not a fetchable
            resource — it is not guaranteed to resolve to a document.
          enum:
            - https://openccu-loom.dev/errors/validation
            - https://openccu-loom.dev/errors/not_found
            - https://openccu-loom.dev/errors/conflict
            - https://openccu-loom.dev/errors/unauthorized
            - https://openccu-loom.dev/errors/forbidden
            - https://openccu-loom.dev/errors/unsupported
            - https://openccu-loom.dev/errors/rate_limited
            - https://openccu-loom.dev/errors/internal
            - https://openccu-loom.dev/errors/bad_request
            - https://openccu-loom.dev/errors/service_unready
            - https://openccu-loom.dev/errors/upstream_unavailable
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        instance: { type: string }
        code:
          type: string
          description: Short tag (also surfaced as the `X-Problem-Code` response header).
          enum:
            - validation
            - not_found
            - conflict
            - unauthorized
            - forbidden
            - unsupported
            - rate_limited
            - internal
            - bad_request
            - service_unready
            - upstream_unavailable
        errors:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              reason: { type: string }
              min: {}
              max: {}
...
    DataPointSummary:
      type: object
      required: [parameter, observed, operations]
      properties:
        parameter: { type: string }
        parameter_label: { type: string }
        value: {}
        observed: { type: boolean }
        modified_at:
          type: string
          format: date-time
          description: |
            When the value last changed on the CCU. RFC3339Nano,
            UTC. Equivalent to `last_changed_at` (kept for
            cache-friendly naming alongside `last_seen_at`).
        source:
          type: string
          enum: [unobserved, cache, live, stale]
          description: |
            Wire-side lifecycle token. See ADR 0018.
            - unobserved: never seen any value
            - cache: restored from persistent cache at boot, awaiting live event
            - live: last value confirmed by CCU push / fetch_all, connection healthy
            - stale: connection lost; value is the last known good
        last_seen_at:
          type: string
          format: date-time
          description: |
            When the data point was last observed via any push or
            fetch_all event (cyclic info telegrams included). Advances
            even when the value did not change.
        last_changed_at:
          type: string
          format: date-time
          description: |
            When the value actually changed last. Alias of modified_at
            kept under the cache-friendly name.
        value_age_seconds:
          type: integer
          format: int64
          description: |
            Pre-computed seconds between last_seen_at and the response
            time, so the browser does not need to parse the timestamp
            on every render.
        operations:
          type: object
          required: [read, write, event]
          properties:
            read: { type: boolean }
            write: { type: boolean }
            event: { type: boolean }
        control:
          type: string
          description: |
            CCU paramset descriptor's CONTROL attribute of the form
            WIDGET_FAMILY.SLOT (e.g. "HEATING_CONTROL_HMIP.SETPOINT").
            Drives the SPA's CONTROL-aware widget resolver. See
            docs/ui/control-widget-concept.md. Empty when the
            descriptor carries no CONTROL.
        type:
          type: string
          description: |
            CCU descriptor TYPE (BOOL, INTEGER, FLOAT, ENUM, STRING, ACTION).
            Surfaced so the SPA can pick the right widget primitive
            without re-reading the paramset descriptor.
        value_list:
          type: array
          items: { type: string }
          description: |
            Ordered enum labels for ENUM-typed parameters
            (e.g. ["BLACK","BLUE","GREEN","TURQUOISE","RED","PURPLE",
            "YELLOW","WHITE"] on HmIP-BSL COLOR). Empty for non-ENUM.
        unit:
          type: string
          description: |
            Descriptor UNIT string ("°C", "% rF", "lx", "Wh", ...)
            as the CCU declares it. Empty when none.
        min: {}
        max: {}
        default: {}
        ui_hint:
          type: object
          description: |
            Daemon-computed UI classification envelope. The SPA's
            AutoTile composer reads the fields verbatim — no JS-side
            classification runs. See docs/ui/auto-tile-concept.md.
            Resolution order in the backend: parameter substring →
            ENUM-shape rules → unit table → type fallback. Never
            empty.
          required: [icon, semantic]
          properties:
            icon:
              type: string
              description: Material Design Icons token, e.g. "mdi:thermometer".
            semantic:
              type: string
              description: |
                Stable token used by the composer to group readouts
                of the same physical quantity into one bucket
                (e.g. "temperature", "particulate", "motion").
            state_color_rule:
              type: string
              description: |
                Named threshold rule the SPA looks up in its rule
                registry to colour the readout (e.g. "temp_heat",
                "alarm_active", "signal_weak"). Empty when no rule.
...
    ValuesBatchRequest:
      type: object
      required: [queries]
      properties:
        queries:
          type: array
          maxItems: 1000
          items:
            type: object
            required: [address, channel, parameter]
            properties:
              address: { type: string }
              channel: { type: integer, minimum: 0 }
              parameter: { type: string }
    ValuesBatchResponse:
      type: object
      required: [results]
      properties:
        results:
          type: array
          items:
            type: object
            required: [address, channel, parameter]
            properties:
              address: { type: string }
              channel: { type: integer, minimum: 0 }
              parameter: { type: string }
              summary: { $ref: "#/components/schemas/DataPointSummary" }
              error:
                type: string
                description: |
                  Per-query failure. Set when `summary` is absent.
                  Stable values: `device not found`,
                  `channel not found`, `parameter not found`,
                  `parameter hidden`.
...